requirements.txt
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cpu
numpy
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
torchvision==0.15.2+cpu
onnxruntime==1.15.1
numpy==1.26.0
pandas==2.2.2
Pillow==9.5.0
opencv-python-headless==4.8.0.74
matplotlib==3.8.1
albumentations==1.3.0

# Web / API
fastapi==0.100.0
uvicorn[standard]==0.23.1
python-multipart==0.0.6

# Other
pyyaml==6.0
scikit-learn==1.3.0
tqdm==4.66.1
streamlit==1.25.0
jupyterlab==4.0.0
